requirements.txt
gitpython
pygerrit2
